lab2_tppo/client/requirements.txt
python==3.9.13
python-socketio==5.7.2
requests==2.28.1